chore: extracts input component
