+++ website/components/FinalForm/index.ts
@@ -1,0 +1,1 @@
+export { default as FFInput } from './FFInput'
